Code Cleanups and fix some bugs with the UI redrawing

+++ ChessAndroid/app/src/main/cpp/chess/algorithm/MoveGen.h
@@ -80,8 +80,8 @@ inline Move parseMove(Board &board, const std::string &str)
 		|| str[0] > 'h' || str[0] < 'a'
 		|| str[2] > 'h' || str[2] < 'a') return {};
 
-    const byte from = toSquare(str[0] - 'a', str[1] - '1');
-    const byte to = toSquare(str[2] - 'a', str[3] - '1');	
+    const byte from = toSquare(static_cast<byte>(str[0] - 'a'), static_cast<byte>(str[1] - '1'));
+    const byte to = toSquare(static_cast<byte>(str[2] - 'a'), static_cast<byte>(str[3] - '1'));
 
 	const MoveList moveList(board);
